# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mediate Healthcare MR App",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/masters/bulk-dropdowns": {
      "get": {
        "tags": [
          "Masters"
        ],
        "summary": "Get all common dropdowns for offline caching",
        "description": "Fetch bundled specializations, categories, and states for client cache.",
        "operationId": "get_bulk_dropdowns_api_v1_masters_bulk_dropdowns_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterBulkDropdownsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/masters": {
      "get": {
        "tags": [
          "Masters"
        ],
        "summary": "List generic master items",
        "description": "Fetch list of business master items.",
        "operationId": "list_master_items_api_v1_masters_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Item type: SPECIALIZATION, CUSTOMER_CATEGORY, VISIT_PRIORITY",
              "title": "Type"
            },
            "description": "Item type: SPECIALIZATION, CUSTOMER_CATEGORY, VISIT_PRIORITY"
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Filter only active items",
              "default": true,
              "title": "Active Only"
            },
            "description": "Filter only active items"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MasterItemResponse"
                  },
                  "title": "Response List Master Items Api V1 Masters Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Masters"
        ],
        "summary": "Create master item (Admin)",
        "description": "Create a new master item.",
        "operationId": "create_master_item_api_v1_masters_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasterItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/masters/{item_id}": {
      "put": {
        "tags": [
          "Masters"
        ],
        "summary": "Update master item (Admin)",
        "description": "Update existing master item.",
        "operationId": "update_master_item_api_v1_masters__item_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasterItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/masters/states": {
      "get": {
        "tags": [
          "Masters"
        ],
        "summary": "List states",
        "description": "List states.",
        "operationId": "list_states_api_v1_masters_states_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StateResponse"
                  },
                  "title": "Response List States Api V1 Masters States Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Masters"
        ],
        "summary": "Create state (Admin)",
        "description": "Create a state.",
        "operationId": "create_state_api_v1_masters_states_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StateCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/masters/cities": {
      "get": {
        "tags": [
          "Masters"
        ],
        "summary": "List cities by state",
        "description": "List cities.",
        "operationId": "list_cities_api_v1_masters_cities_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "state_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State Id"
            }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CityResponse"
                  },
                  "title": "Response List Cities Api V1 Masters Cities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Masters"
        ],
        "summary": "Create city (Admin)",
        "description": "Create a city.",
        "operationId": "create_city_api_v1_masters_cities_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CityCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/masters/areas": {
      "get": {
        "tags": [
          "Masters"
        ],
        "summary": "List areas by city",
        "description": "List areas.",
        "operationId": "list_areas_api_v1_masters_areas_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "city_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City Id"
            }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AreaResponse"
                  },
                  "title": "Response List Areas Api V1 Masters Areas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Masters"
        ],
        "summary": "Create area (Admin)",
        "description": "Create an area.",
        "operationId": "create_area_api_v1_masters_areas_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AreaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AreaCreate": {
        "properties": {
          "city_id": {
            "type": "integer",
            "title": "City Id"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "pincode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Pincode"
          }
        },
        "type": "object",
        "required": [
          "city_id",
          "name"
        ],
        "title": "AreaCreate",
        "description": "Payload to create an Area."
      },
      "AreaResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "city_id": {
            "type": "integer",
            "title": "City Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "pincode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pincode"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "city_id",
          "name",
          "is_active"
        ],
        "title": "AreaResponse",
        "description": "Area / Locality response."
      },
...
      "CityCreate": {
        "properties": {
          "state_id": {
            "type": "integer",
            "title": "State Id"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "state_id",
          "name"
        ],
        "title": "CityCreate",
        "description": "Payload to create a City."
      },
      "CityResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "state_id": {
            "type": "integer",
            "title": "State Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state_id",
          "name",
          "is_active"
        ],
        "title": "CityResponse",
        "description": "City / District response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MasterBulkDropdownsResponse": {
        "properties": {
          "specializations": {
            "items": {
              "$ref": "#/components/schemas/MasterItemResponse"
            },
            "type": "array",
            "title": "Specializations"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/MasterItemResponse"
            },
            "type": "array",
            "title": "Categories"
          },
          "priorities": {
            "items": {
              "$ref": "#/components/schemas/MasterItemResponse"
            },
            "type": "array",
            "title": "Priorities"
          },
          "states": {
            "items": {
              "$ref": "#/components/schemas/StateResponse"
            },
            "type": "array",
            "title": "States"
          }
        },
        "type": "object",
        "required": [
          "specializations",
          "categories",
          "priorities",
          "states"
        ],
        "title": "MasterBulkDropdownsResponse",
        "description": "Combined masters response for offline cache & dropdowns."
      },
      "MasterItemCreate": {
        "properties": {
          "type": {
            "type": "string",
            "maxLength": 50,
            "title": "Type",
            "description": "e.g. SPECIALIZATION, CUSTOMER_CATEGORY"
          },
          "code": {
            "type": "string",
            "maxLength": 50,
            "title": "Code",
            "description": "Unique code within type"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "type",
          "code",
          "name"
        ],
        "title": "MasterItemCreate",
        "description": "Payload to create a generic master item."
      },
      "MasterItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "code",
          "name",
          "sort_order",
          "is_active",
          "created_at"
        ],
        "title": "MasterItemResponse",
        "description": "Generic business master item response."
      },
      "MasterItemUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "MasterItemUpdate",
        "description": "Payload to update a generic master item."
      },
...
      "StateCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "code": {
            "type": "string",
            "maxLength": 10,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "name",
          "code"
        ],
        "title": "StateCreate",
        "description": "Payload to create a State."
      },
      "StateResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "code",
          "is_active"
        ],
        "title": "StateResponse",
        "description": "State / Province response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}